Restructure into catkin_workspace directory structure, add visualization
