+++ .gitignore
@@ -1,0 +1,3 @@
+**/build/
+**/devel/
+*.catkin_workspace
