+++ catkin_ws/src/baymax_visualization/launch/.ipynb_checkpoints/dofbot_moveit-checkpoint.launch
@@ -1,0 +1,9 @@
+<launch>
+    <arg name="model" default="dofbot.urdf"/>
+    <arg name="gui" default="false"/>
+    <param name="robot_description" command="$(find xacro)/xacro --inorder $(find dofbot_moveit)/urdf/$(arg model)"/>
+    <param name="use_gui" value="$(arg gui)"/>
+    <node name="joint_state_publisher_gui" pkg="joint_state_publisher_gui" type="joint_state_publisher_gui"/>
+    <node name="robot_state_publisher" pkg="robot_state_publisher" type="robot_state_publisher"/>
+    <node name="rviz" pkg="rviz" type="rviz" args="-d $(find dofbot_moveit)/rviz/dofbot.rviz" required="true"/>
+</launch>
