+++ catkin_ws/src/baymax_visualization/launch/.ipynb_checkpoints/fk-checkpoint.launch
@@ -1,0 +1,3 @@
+<launch>
+    <node name="fk" pkg="dofbot_moveit" type="dofbot_kinematics_fk" output="screen"/>
+</launch>
