+++ catkin_ws/src/baymax_visualization/launch/.ipynb_checkpoints/ik-checkpoint.launch
@@ -1,0 +1,3 @@
+<launch>
+    <node name="dofbot_ik" pkg="dofbot_moveit" type="dofbot_kinematics_ik" output="screen"/>
+</launch>
