+++ catkin_ws/src/baymax_visualization/launch/dofbot_moveit.launch
@@ -1,0 +1,8 @@
+<launch>
+    <arg name="model" default="dofbot.urdf"/>
+    <arg name="gui" default="false"/>
+    <param name="robot_description" command="$(find xacro)/xacro --inorder $(find baymax_visualization)/urdf/$(arg model)"/>
+    <param name="use_gui" value="$(arg gui)"/>
+    <node name="robot_state_publisher" pkg="robot_state_publisher" type="robot_state_publisher"/>
+    <node name="rviz" pkg="rviz" type="rviz" args="-d $(find baymax_visualization)/rviz/dofbot.rviz" required="true"/>
+</launch>
